refactor(cli-tools): consolidate OXC config merger
